requirements.txt
torch
transformers
open3d
Pillow
matplotlib
numpy